+++ app/src/main/java/com/piotr/practicepad/exerciseList/ExerciseSetEntity.kt
@@ -5,7 +5,7 @@ import com.piotr.practicepad.exercise.Exercise
 data class ExerciseSet(
     val id: Int = 0,
     val title: String = "",
-    val tempo: Int = 0,
+    val tempo: Long = 0L,
     val exercises: List<Exercise> = listOf()
 ) {
     fun shouldStartNextExercise(position: Int): Boolean = position < exercises.size
