+++ app/src/main/java/com/piotr/practicepad/exerciseSetDetail/ExerciseSetDetailsState.kt
@@ -4,6 +4,6 @@ import com.piotr.practicepad.exercise.Exercise
 
 data class ExerciseSetDetailsState(
     val name: String = "",
-    val tempo: Int = 0,
+    val tempo: Long = 0,
     val exerciseDetailsList: MutableList<Exercise> = mutableListOf()
 )
